updated professor miolane comments
